fix pascal and kotlin code for longint

+++ tests/sum_two_numbers/kotlin_sol.kt
@@ -1,5 +1,5 @@
 fun main() {
-    val a = readLine()!!.toInt()
-    val b = readLine()!!.toInt()
+    val a = readLine()!!.toLong()
+    val b = readLine()!!.toLong()
     println(a + b)
 }
